Create bearing-washer.stl
The official bearing guide and fender washer solution have one thing that is not convenient: the bearing itself cannot be position restricted unless it will not be able to rotate. This bearing washer solves this problem by increase the thickness close to the hole. Since the 608 washer have a ring around its hole, this ring can then be restricted by touching this fender properly. At the same time, the other part of the bearing can still rotate freely.